requirements.txt
pandas>=2.0,<3
pyyaml>=6.0
requests>=2.28,<3
matplotlib>=3.7,<4
numpy>=1.26,<3
scikit-learn>=1.3,<2
